+++ lib/main.dart
@@ -1,9 +1,9 @@
 import 'package:diary_demo_app/domain/life_balance_wheel.dart';
 import 'package:diary_demo_app/domain/week_goal.dart';
 import 'package:diary_demo_app/domain/week_review.dart';
-import 'package:diary_demo_app/week_widget.dart';
-import 'package:diary_demo_app/life_balance_wheel_widget.dart';
-import 'package:diary_demo_app/setting_widget.dart';
+import 'package:diary_demo_app/widget/week_widget.dart';
+import 'package:diary_demo_app/widget/life_balance_wheel_widget.dart';
+import 'package:diary_demo_app/widget/setting_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
